+++ src/cross-client.ts
@@ -28,6 +28,7 @@ import {
   TransactionInstruction,
   ConfirmOptions,
   SYSVAR_CLOCK_PUBKEY,
+  ComputeBudgetProgram,
 } from "@solana/web3.js";
 import * as splToken from "@solana/spl-token";
 import { PublicKey as PublicKeyZstd } from "zeta-solana-web3";
@@ -710,6 +711,10 @@ export class CrossClient {
   ): Promise<TransactionSignature> {
     let tx = new Transaction();
 
+    if ((await this.hasReferrerAccounts()) == true) {
+      throw Error("User already has a referrer account!");
+    }
+
     if (this._accountManager === null) {
       console.log(
         "User has no cross margin account manager. Creating account manager..."
@@ -1244,6 +1249,71 @@ export class CrossClient {
     return txId;
   }
 
+  public async hasReferrerAccounts(): Promise<boolean> {
+    // Firstly, close all referrer accounts
+    let referrerPubkeyAccountAddress = utils.getReferrerPubkeyAccount(
+      Exchange.programId,
+      this.publicKey
+    )[0];
+    let referrerPubkeyAccount =
+      await Exchange.program.account.referrerPubkeyAccount.fetchNullable(
+        referrerPubkeyAccountAddress
+      );
+    return referrerPubkeyAccount != null;
+  }
+
+  public async remakeReferrerAccounts(
+    id: string
+  ): Promise<TransactionSignature> {
+    this.delegatedCheck();
+
+    let tx = new Transaction();
+
+    // Firstly, close all referrer accounts
+    if ((await this.hasReferrerAccounts()) == true) {
+      let referrerPubkeyAccountAddress = utils.getReferrerPubkeyAccount(
+        Exchange.programId,
+        this.publicKey
+      )[0];
+      let referrerPubkeyAccount =
+        await Exchange.program.account.referrerPubkeyAccount.fetch(
+          referrerPubkeyAccountAddress
+        );
+      let id = Buffer.from(referrerPubkeyAccount.referrerId).toString();
+      let referrerIdAccountAddress = utils.getReferrerIdAccount(
+        Exchange.programId,
+        id
+      )[0];
+
+      tx.add(
+        instructions.closeReferrerAccountsIx(
+          this.publicKey,
+          referrerIdAccountAddress,
+          referrerPubkeyAccountAddress
+        )
+      );
+    }
+
+    // Secondly, make new ones with the provided ID
+    tx.add(
+      instructions.initializeReferrerAccountsIx(
+        id,
+        this.publicKey,
+        utils.getReferrerIdAccount(Exchange.programId, id)[0],
+        utils.getReferrerPubkeyAccount(Exchange.programId, this.publicKey)[0]
+      )
+    );
+
+    return await utils.processTransaction(
+      this._provider,
+      tx,
+      undefined,
+      undefined,
+      undefined,
+      this._useVersionedTxs ? utils.getZetaLutArr() : undefined
+    );
+  }
+
   public async closeReferrerAccounts(): Promise<TransactionSignature> {
     this.delegatedCheck();
 
@@ -1301,7 +1371,11 @@ export class CrossClient {
     options: types.OrderOptions = types.defaultOrderOptions(),
     preIxs: TransactionInstruction[] = []
   ): Promise<TransactionSignature> {
-    let tx = new Transaction();
+    let tx = new Transaction().add(
+      ComputeBudgetProgram.setComputeUnitLimit({
+        units: 300_000,
+      })
+    );
     let assetIndex = assetToIndex(asset);
     let market = Exchange.getPerpMarket(asset);
     let openOrdersPda = null;
@@ -1914,6 +1988,32 @@ export class CrossClient {
     return txSig;
   }
 
+  public async chooseAirdropCommunity(
+    community: types.AirdropCommunity
+  ): Promise<TransactionSignature> {
+    this.delegatedCheck();
+    let tx = new Transaction();
+
+    tx.add(
+      instructions.chooseAirdropCommunityIx(
+        community,
+        this.accountManagerAddress,
+        this._provider.wallet.publicKey
+      )
+    );
+
+    return await utils.processTransaction(
+      this._provider,
+      tx,
+      undefined,
+      undefined,
+      undefined,
+      this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+      undefined,
+      this._txRetryAmount
+    );
+  }
+
   public async editDelegatedPubkey(
     delegatedPubkey: PublicKey
   ): Promise<TransactionSignature> {
@@ -2328,7 +2428,11 @@ export class CrossClient {
   ): Promise<TransactionSignature> {
     let market = Exchange.getPerpMarket(asset);
     let assetIndex = assetToIndex(asset);
-    let tx = new Transaction();
+    let tx = new Transaction().add(
+      ComputeBudgetProgram.setComputeUnitLimit({
+        units: 300_000,
+      })
+    );
     tx.add(
       instructions.cancelOrderIx(
         asset,
@@ -2395,7 +2499,11 @@ export class CrossClient {
   ): Promise<TransactionSignature> {
     let market = Exchange.getPerpMarket(asset);
     let assetIndex = assetToIndex(asset);
-    let tx = new Transaction();
+    let tx = new Transaction().add(
+      ComputeBudgetProgram.setComputeUnitLimit({
+        units: 300_000,
+      })
+    );
     tx.add(
       instructions.cancelOrderByClientOrderIdIx(
         asset,
@@ -2462,7 +2570,11 @@ export class CrossClient {
     newOrderSide: types.Side,
     newOptions: types.OrderOptions = types.defaultOrderOptions()
   ): Promise<TransactionSignature> {
-    let tx = new Transaction();
+    let tx = new Transaction().add(
+      ComputeBudgetProgram.setComputeUnitLimit({
+        units: 300_000,
+      })
+    );
     let market = Exchange.getPerpMarket(asset);
     let assetIndex = assetToIndex(asset);
     tx.add(