+++ src/utils.ts
@@ -51,6 +51,7 @@ import cloneDeep from "lodash.clonedeep";
 import * as os from "os";
 import { OpenOrders, _OPEN_ORDERS_LAYOUT_V2 } from "./serum/market";
 import { HttpProvider } from "@bloxroute/solana-trader-client-ts";
+import axios from "axios";
 
 export function getNativeTickSize(asset: Asset): number {
   return Exchange.state.tickSizes[assets.assetToIndex(asset)];
@@ -495,7 +496,7 @@ export function getReferrerIdAccount(
   return anchor.web3.PublicKey.findProgramAddressSync(
     [
       Buffer.from(anchor.utils.bytes.utf8.encode("referrer-id-account")),
-      Buffer.from(id),
+      Buffer.from(id.replace(/[\0]+$/g, "")),
     ],
     programId
   );
@@ -1058,6 +1059,122 @@ export async function processTransactionBloxroute(
   }
 }
 
+export async function processTransactionJito(
+  provider: anchor.AnchorProvider,
+  tx: Transaction,
+  signers?: Array<Signer>,
+  opts?: ConfirmOptions,
+  lutAccs?: AddressLookupTableAccount[],
+  blockhash?: { blockhash: string; lastValidBlockHeight: number }
+): Promise<TransactionSignature> {
+  if (Exchange.jitoTip == 0) {
+    throw Error("Jito bundle tip has not been set.");
+  }
+
+  if (Exchange.priorityFee != 0) {
+    tx.instructions.unshift(
+      ComputeBudgetProgram.setComputeUnitPrice({
+        microLamports: Math.round(Exchange.priorityFee),
+      })
+    );
+  }
+
+  tx.instructions.push(
+    SystemProgram.transfer({
+      fromPubkey: Exchange.provider.publicKey,
+      toPubkey: new PublicKey(
+        "DttWaMuVvTiduZRnguLF7jNxTgiMBZ1hyAumKUiL2KRL" // Jito tip account
+      ),
+      lamports: Exchange.jitoTip, // tip
+    })
+  );
+
+  let recentBlockhash =
+    blockhash ??
+    (await provider.connection.getLatestBlockhash(
+      Exchange.blockhashCommitment
+    ));
+
+  let vTx: VersionedTransaction = new VersionedTransaction(
+    new TransactionMessage({
+      payerKey: provider.wallet.publicKey,
+      recentBlockhash: recentBlockhash.blockhash,
+      instructions: tx.instructions,
+    }).compileToV0Message(lutAccs)
+  );
+
+  vTx.sign(
+    (signers ?? [])
+      .filter((s) => s !== undefined)
+      .map((kp) => {
+        return kp;
+      })
+  );
+  vTx = (await provider.wallet.signTransaction(vTx)) as VersionedTransaction;
+
+  let rawTx = vTx.serialize();
+
+  const encodedTx = bs58.encode(rawTx);
+  const jitoURL = "https://mainnet.block-engine.jito.wtf/api/v1/transactions";
+  const payload = {
+    jsonrpc: "2.0",
+    id: 1,
+    method: "sendTransaction",
+    params: [encodedTx],
+  };
+
+  let txOpts = opts || commitmentConfig(provider.connection.commitment);
+  let txSig: string;
+  try {
+    const response = await axios.post(jitoURL, payload, {
+      headers: { "Content-Type": "application/json" },
+    });
+    txSig = response.data.result;
+  } catch (error) {
+    console.error("Error:", error);
+    throw new Error("Jito Bundle Error: cannot send.");
+  }
+
+  if (Exchange.skipRpcConfirmation) {
+    return txSig;
+  }
+
+  let currentBlockHeight = await provider.connection.getBlockHeight(
+    provider.connection.commitment
+  );
+
+  while (currentBlockHeight < recentBlockhash.lastValidBlockHeight) {
+    // Keep resending to maximise the chance of confirmation
+    await provider.connection.sendRawTransaction(rawTx, {
+      skipPreflight: true,
+      preflightCommitment: provider.connection.commitment,
+    });
+
+    let status = await provider.connection.getSignatureStatus(txSig);
+    currentBlockHeight = await provider.connection.getBlockHeight(
+      provider.connection.commitment
+    );
+    if (status.value != null) {
+      if (status.value.err != null) {
+        // Gets caught and parsed in the later catch
+        let err = parseInt(status.value.err["InstructionError"][1]["Custom"]);
+        let parsedErr = parseError(err);
+        throw parsedErr;
+      }
+      if (
+        txConfirmationCheck(
+          txOpts.commitment ? txOpts.commitment.toString() : "confirmed",
+          status.value.confirmationStatus.toString()
+        )
+      ) {
+        return txSig;
+      }
+    }
+    await sleep(500); // Don't spam the RPC
+  }
+  throw Error(`Transaction ${txSig} was not confirmed`);
+}
+
 export async function processTransaction(
   provider: anchor.AnchorProvider,
   tx: Transaction,
@@ -1066,9 +1183,19 @@ export async function processTransaction(
   useLedger: boolean = false,
   lutAccs?: AddressLookupTableAccount[],
   blockhash?: { blockhash: string; lastValidBlockHeight: number },
-  retries?: number,
-  skipConfirmation?: boolean
+  retries?: number
 ): Promise<TransactionSignature> {
+  if (Exchange.useJitoBundle) {
+    return processTransactionJito(
+      provider,
+      tx,
+      signers,
+      opts,
+      lutAccs,
+      blockhash
+    );
+  }
+
   if (Exchange.httpProvider) {
     let txSig = await processTransactionBloxroute(
       Exchange.httpProvider,
@@ -1077,7 +1204,7 @@ export async function processTransaction(
       Exchange.tipMultiplier * Exchange.priorityFee + 1050,
       blockhash,
       retries,
-      skipConfirmation
+      Exchange.skipRpcConfirmation
     );
     return txSig;
   }
@@ -1157,13 +1284,20 @@ export async function processTransaction(
 
       txSig = await provider.connection.sendRawTransaction(rawTx, {
         skipPreflight: true,
+        preflightCommitment: provider.connection.commitment,
       });
 
       // Poll the tx confirmation for N seconds
       // Polling is more reliable than websockets using confirmTransaction()
       let currentBlockHeight = 0;
-      if (!skipConfirmation) {
+      if (!Exchange.skipRpcConfirmation) {
         while (currentBlockHeight < recentBlockhash.lastValidBlockHeight) {
+          // Keep resending to maximise the chance of confirmation
+          await provider.connection.sendRawTransaction(rawTx, {
+            skipPreflight: true,
+            preflightCommitment: provider.connection.commitment,
+          });
+
           let status = await provider.connection.getSignatureStatus(txSig);
           currentBlockHeight = await provider.connection.getBlockHeight(
             provider.connection.commitment
@@ -1185,7 +1319,7 @@ export async function processTransaction(
               return txSig;
             }
           }
-          await sleep(1500); // Don't spam the RPC
+          await sleep(500); // Don't spam the RPC
         }
         throw Error(`Transaction ${txSig} was not confirmed`);
       } else {
@@ -1437,13 +1571,31 @@ export async function cleanZetaMarketHalted(asset: Asset) {
  */
 export async function crankMarket(
   asset: Asset,
-  openOrdersToMargin?: Map<PublicKey, PublicKey>,
+  openOrdersToMargin?: Map<string, PublicKey>,
   crankLimit?: number
 ): Promise<boolean> {
+  let ix = await createCrankMarketIx(asset, openOrdersToMargin, crankLimit);
+  if (ix == null) return true;
+  let tx = new Transaction()
+    .add(
+      ComputeBudgetProgram.setComputeUnitLimit({
+        units: 250_000,
+      })
+    )
+    .add(ix);
+  await processTransaction(Exchange.provider, tx);
+  return false;
+}
+
+export async function createCrankMarketIx(
+  asset: Asset,
+  openOrdersToMargin?: Map<string, PublicKey>,
+  crankLimit?: number
+): Promise<TransactionInstruction | null> {
   let market = Exchange.getPerpMarket(asset);
   let eventQueue = await market.serumMarket.loadEventQueue(Exchange.connection);
   if (eventQueue.length == 0) {
-    return true;
+    return null;
   }
   const openOrdersSet = new Set();
   // We pass in a couple of extra accounts for perps so the limit is lower
@@ -1467,15 +1619,20 @@ export async function crankMarket(
 
   let remainingAccounts: any[] = new Array(uniqueOpenOrders.length * 2);
 
-  // TODO test support for both crossmargin and marginaccounts
   await Promise.all(
     uniqueOpenOrders.map(async (openOrders, index) => {
       let marginAccount: PublicKey;
-      if (openOrdersToMargin && !openOrdersToMargin.has(openOrders)) {
+      if (
+        openOrdersToMargin &&
+        !openOrdersToMargin.has(openOrders.toBase58())
+      ) {
         marginAccount = await getAccountFromOpenOrders(openOrders, asset);
-        openOrdersToMargin.set(openOrders, marginAccount);
-      } else if (openOrdersToMargin && openOrdersToMargin.has(openOrders)) {
-        marginAccount = openOrdersToMargin.get(openOrders);
+        openOrdersToMargin.set(openOrders.toBase58(), marginAccount);
+      } else if (
+        openOrdersToMargin &&
+        openOrdersToMargin.has(openOrders.toBase58())
+      ) {
+        marginAccount = openOrdersToMargin.get(openOrders.toBase58());
       } else {
         marginAccount = await getAccountFromOpenOrders(openOrders, asset);
       }
@@ -1494,23 +1651,13 @@ export async function crankMarket(
     })
   );
 
-  let tx = new Transaction()
-    .add(
-      ComputeBudgetProgram.setComputeUnitLimit({
-        units: 250_000,
-      })
-    )
-    .add(
-      instructions.crankMarketIx(
-        asset,
-        market.address,
-        market.serumMarket.eventQueueAddress,
-        constants.DEX_PID[Exchange.network],
-        remainingAccounts
-      )
-    );
-  await processTransaction(Exchange.provider, tx);
-  return false;
+  return instructions.crankMarketIx(
+    asset,
+    market.address,
+    market.serumMarket.eventQueueAddress,
+    constants.DEX_PID[Exchange.network],
+    remainingAccounts
+  );
 }
 
 /*
@@ -2072,6 +2219,16 @@ export function median(arr: number[]): number | undefined {
   return s.length % 2 === 0 ? (s[mid - 1] + s[mid]) / 2 : s[mid];
 }
 
+export async function isAffiliateCodeAvailable(code: string): Promise<boolean> {
+  let referrerIdAddress = getReferrerIdAccount(Exchange.programId, code)[0];
+  let referrerIdAccount =
+    await Exchange.program.account.referrerIdAccount.fetchNullable(
+      referrerIdAddress
+    );
+
+  return referrerIdAccount == null;
+}
+
 export const checkLiquidity = (
   size: number,
   asset: Asset,
@@ -2341,6 +2498,14 @@ export function calculateTakeTriggerOrderExecutionPrice(
     );
   }
   return executionPrice;
+}
+
+export function getFeeTier(accountType: constants.MarginAccountType): number {
+  let tier = constants.ACCOUNT_TYPE_TO_FEE_TIER_MAP[accountType];
+  if (tier == undefined) {
+    return 0;
+  }
+  return tier;
 }
 
 export function getFeeBps(